Don't reverse steps on migration failure if backend supports transactional DDL
For fully transactional backends reversing migrations steps is unecessary:
everything gets rolled back at the end anyway.
If the reverse steps also fail the script output contains multiple tracebacks
which is unexpected and hard to interpret.

For MySQL (and any other backends without a transactional DDL),
we retain the existing behaviour of running the reverse steps to try to undo
any steps applied so far. This will hopefully leave the database in a
consistent state.