tests: replace `@with_migrations` decorator with equivalent context manager
This has been causing problems because pytest fixtures don't work well with
function decorators.

If `with_migrations` doesn't correctly wrap the decorator using `functools.wraps`,
pytest fixtures cannot be used by the decorated test function.

But if we do use `functools.wraps`, pytest fixtures assumes the extra function
argument is a fixture name and then fails either because it conflicts with the
builtin `tmpdir` fixture or, if we rename it, because such a fixture doesn't
exist.