Fix access to StepCollector from imported yoyo.migrations.step
Fixes https://todo.sr.ht/~olly/yoyo/79

The old logic relied on a string comparison of the module path. This broke when
Python's importlib switched to using absolute paths.

It's safer to simply look for a ``__yoyo_collector__`` attribute in the caller
frame's namespace, which is injected by Migration.load. I also chose to rename
this attribute from 'collector' to __yoyo_collector__ to reduce
the risk of collision with a user defined variable.